Helloworld/app/src/androidTest/java/com/framgia/common/UtilTest.java
package com.framgia.common;

import android.content.Context;
import android.support.test.filters.SdkSuppress;
import android.test.suitebuilder.annotation.SmallTest;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.io.FileOutputStream;

import static org.junit.Assert.fail;
import static org.mockito.Matchers.any;
import static org.mockito.Matchers.anyInt;
import static org.mockito.Matchers.anyString;
import static org.mockito.Mockito.atLeast;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

/**
 * Created by FRAMGIA\pham.van.khac on 22/01/2016.
 */
public class UtilTest {
    @Mock
    Context context;

    @Mock
    FileOutputStream fileOutputStream;

    @Before
    public void init() {
        MockitoAnnotations.initMocks(this);
    }

    @Test
    @SmallTest
    @SdkSuppress(minSdkVersion = 18)
    public void writeShouldWriteTwiceToFileSystem() {
        try {
            when(context.openFileOutput(anyString(), anyInt())).thenReturn(fileOutputStream);
            Util.writeConfiguration(context);
            verify(context, times(1)).openFileOutput(anyString(), anyInt());
            verify(fileOutputStream, atLeast(2)).write(any(byte[].class));

        } catch (Exception e) {
            e.printStackTrace();
            fail();
        }
    }
}
